DPL: have separate include for RootMessageContext
This means that we can now spot all the messages which are
serialised via ROOT by simply removing

 #include "Framework/RootMessageContext.h"

from DataAllocator.h
