+++ Framework/Core/include/Framework/MessageContext.h
@@ -17,7 +17,6 @@
 #include "Framework/RouteState.h"
 #include "Framework/RoutingIndices.h"
 #include "Framework/RuntimeError.h"
-#include "Framework/TMessageSerializer.h"
 #include "Framework/DataProcessingHeader.h"
 #include "Framework/TypeTraits.h"
 
@@ -39,6 +38,15 @@
 
 namespace o2::framework
 {
+
+template <typename T, typename = void>
+struct enable_root_serialization : std::false_type {
+  using debug_type = T;
+};
+
+template <typename T, typename = void>
+struct root_serializer : std::false_type {
+};
 
 struct Output;
 
@@ -350,55 +358,6 @@ class MessageContext
 
    private:
     value_type mValue;
-  };
-
-  /// RootSerializedObject keeps ownership to an object which can be Root-serialized
-  /// TODO: this should maybe be a separate header file to avoid including TMessageSerializer
-  /// in this header file, but we can always change this without affecting to much code.
-  template <typename T>
-  class RootSerializedObject : public ContextObject
-  {
-   public:
-    // Note: we strictly require the type to implement the ROOT ClassDef interface in order to be
-    // able to check for the existence of the dirctionary for this type. Could be dropped if any
-    // use case for a type having the dictionary at runtime pops up
-    static_assert(has_root_dictionary<T>::value == true, "unconsistent type: needs to implement ROOT ClassDef interface");
-    using value_type = T;
-    /// default constructor forbidden, object alwasy has to control messages
-    RootSerializedObject() = delete;
-    /// constructor taking header message by move and creating the object from variadic argument list
-    template <typename ContextType, typename... Args>
-    RootSerializedObject(ContextType* context, fair::mq::MessagePtr&& headerMsg, RouteIndex routeIndex, Args&&... args)
-      : ContextObject(std::forward<fair::mq::MessagePtr>(headerMsg), routeIndex)
-    {
-      mObject = std::make_unique<value_type>(std::forward<Args>(args)...);
-      mPayloadMsg = context->proxy().createOutputMessage(routeIndex);
-    }
-    ~RootSerializedObject() override = default;
-
-    /// @brief Finalize object and return parts by move
-    /// This retrieves the actual message from the vector object and moves it to the parts
-    fair::mq::Parts finalize() final
-    {
-      assert(mParts.Size() == 1);
-      TMessageSerializer::Serialize(*mPayloadMsg, mObject.get(), nullptr);
-      mParts.AddPart(std::move(mPayloadMsg));
-      return ContextObject::finalize();
-    }
-
-    operator value_type&()
-    {
-      return *mObject;
-    }
-
-    value_type& get()
-    {
-      return *mObject;
-    }
-
-   private:
-    std::unique_ptr<value_type> mObject;
-    fair::mq::MessagePtr mPayloadMsg;
   };
 
   using Messages = std::vector<std::unique_ptr<ContextObject>>;
